demo/requirements.txt
streamlit
streamlit-folium
snowflake-connector-python
snowflake-snowpark-python
pandas
replicate
beautifulsoup4
